mySQL_project2/requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==5.3.0
etelemetry==0.3.1
filelock==3.12.2
fitz==0.0.1.dev2
Flask==2.2.5
httplib2==0.22.0
idna==3.7
importlib-metadata==6.7.0
isodate==0.6.1
itsdangerous==2.1.2
jinja2==3.1.4
looseversion==1.3.0
lxml==5.2.2
MarkupSafe==2.1.5
mysql-connector-python==8.0.33
networkx==2.6.3
nibabel==4.0.2
nipype==1.8.6
numpy==1.21.6
packaging==24.0
pandas==1.1.5
pathlib==1.0.1
protobuf==3.20.3
prov==2.0.1
pydot==2.0.0
PyMuPDF==1.22.5
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyxnat==1.6.2
rdflib==6.3.2
requests==2.31.0
scipy==1.7.3
simplejson==3.19.2
six==1.16.0
traits==6.3.2
typing-extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
